move query hooks from @lib/query -> @lib/hooks/query
