+++ client/src/components/Today/Tasks.tsx
@@ -2,8 +2,8 @@ import Empty from "@/components/Today/Empty";
 import NewActivity from "@/components/activities/NewActivity/NewActivity";
 import Modal from "@/components/utility/Modal/Modal";
 import { filterTagsById } from "@/lib/filter-tags";
+import useTagsQuery from "@/lib/hooks/query/tags/useTagsQuery";
 import modalIds from "@/lib/modal-ids";
-import useTagsQuery from "@/lib/query/tags/useTagsQuery";
 import { useModalState } from "@/lib/state/modal-state";
 import type { ActivityWithIds } from "@type/server/activity.types";
 import Task from "./Task";
