+++ client/src/components/Today/Today.tsx
@@ -9,9 +9,9 @@ import TimelineRows from "@/components/Today/TimelineRows";
 import Calendar from "@/components/utility/Calendar/Calendar";
 import Modal from "@/components/utility/Modal/Modal";
 import { today } from "@/lib/datetime/make-date";
+import useActivitiesQuery from "@/lib/hooks/query/activities/useActivitiesQuery";
 import useHabitsData from "@/lib/hooks/useHabitsData";
 import modalIds from "@/lib/modal-ids";
-import useActivitiesQuery from "@/lib/query/activities/useActivitiesQuery";
 import { useModalState } from "@/lib/state/modal-state";
 import type { TimeWindow } from "@/types/time-window.types";
 import { activityFallsOnDay, isAllDayActivityOnDate } from "@lib/activity";
