+++ client/src/components/activities/ActivityList/useActivityList.ts
@@ -1,5 +1,5 @@
-import useActivitiesQuery from "@/lib/query/activities/useActivitiesQuery";
-import useTagsQuery from "@/lib/query/tags/useTagsQuery";
+import useActivitiesQuery from "@/lib/hooks/query/activities/useActivitiesQuery";
+import useTagsQuery from "@/lib/hooks/query/tags/useTagsQuery";
 
 export default function useActivityList() {
 	const { data: activitiesData } = useActivitiesQuery();
