+++ client/src/components/activities/NewActivity/useNewActivity.ts
@@ -2,9 +2,9 @@ import { queryClient } from "@/lib/query-client";
 import { useModalState } from "@/lib/state/modal-state";
 import type { Datelike } from "@/types/date.types";
 import { hasValidUserId } from "@/types/server/user-id.guards";
+import { useNewActivityMutation } from "@lib/hooks/query/activities/useNewActivityMutation";
 import useAuthentication from "@lib/hooks/useAuthentication";
 import useRouteProps from "@lib/hooks/useRouteProps";
-import { useNewActivityMutation } from "@lib/query/activities/useNewActivityMutation";
 import { useTagSelection } from "@lib/state/selected-tags-state";
 import type { DateTimeField } from "@type/form.types";
 import type { NewActivity } from "@type/server/activity.types";
