+++ client/src/components/habits/DetailedHabit/DetailedHabit.tsx
@@ -1,5 +1,5 @@
 import { createDate } from "@/lib/datetime/make-date";
-import useTagsQuery from "@/lib/query/tags/useTagsQuery";
+import useTagsQuery from "@/lib/hooks/query/tags/useTagsQuery";
 import C from "@/lib/theme/components/Card.style";
 import type { HabitWithIds } from "@/types/server/habit.types";
 import type { PropsWithChildren } from "react";
