+++ client/src/components/habits/NewHabit/useNewHabit.ts
@@ -1,8 +1,8 @@
 import { createDate } from "@/lib/datetime/make-date";
+import { useNewHabitMutation } from "@/lib/hooks/query/habits/useNewHabitMutation";
 import useAuthentication from "@/lib/hooks/useAuthentication";
 import useRouteProps from "@/lib/hooks/useRouteProps";
 import modalIds from "@/lib/modal-ids";
-import { useNewHabitMutation } from "@/lib/query/habits/useNewHabitMutation";
 import { useModalState } from "@/lib/state/modal-state";
 import { useTagSelection } from "@/lib/state/selected-tags-state";
 import type { NewHabit } from "@/types/server/habit.types";
