+++ client/src/lib/hooks/query/activities/useActivitiesQuery.ts
