+++ client/src/lib/hooks/query/habits/useNewHabitEntryMutation.ts
