+++ client/src/lib/hooks/query/habits/useNewHabitMutation.ts
